backend/pyproject.toml
# pyproject.toml

[project]
name = "ai_video_platform"
version = "0.1.0"
description = "AI platform for accelerating video pre-production."
authors = [{ name = "Your Name", email = "your@email.com" }]
requires-python = ">=3.11" # Set to 3.11 based on previous decision
license = { text = "MIT" }
readme = "README.md"
dependencies = [
  "fastapi>=0.115.0,<0.116.0",
  "uvicorn[standard]>=0.34.0,<0.35.0",
  "haystack-ai>=2.0.0,<3.0.0",
  "sqlalchemy>=2.0.35,<2.1.0",
  "pgvector>=0.4.0,<0.5.0",
  "alembic>=1.15.1,<1.16.0",
  "asyncpg>=0.30.0,<0.31.0",                     # Added for PostgreSQL async support
  "redis>=5.2.0,<5.3.0",
  "pydantic-settings>=2.8.0,<2.9.0",
  "pydantic>=2.11.0,<2.12.0",
  "python-jose[cryptography]>=3.4.0,<3.5.0",
  "opentelemetry-api>=1.31.0,<1.33.0",
  "opentelemetry-sdk>=1.31.0,<1.33.0",
  "opentelemetry-exporter-otlp>=1.31.0,<1.33.0",
  "portkey-ai>=1.12.0,<1.13.0",                  # Updated version
  "saq>=0.22.5,<0.23.0",                         # Added SAQ with version 0.22.5+ as specified in requirements
  "supertokens-python>=0.29.1,<0.30.0",          # Added SuperTokens SDK
  "email-validator>=2.2.0,<3.0.0",               # Added for Pydantic EmailStr validation
  "pytest>=8.1.1,<8.2.0",                        # Moved from dev dependencies
  "pytest-mock>=3.14.0,<3.20.0",
  "pytest-asyncio>=0.23.6,<0.24.0",              # Moved from dev dependencies
  "pyyaml>=6.0.0,<7.0.0",                        # Moved from dev dependencies - Required for loading prompts.yaml
]

[project.optional-dependencies]
dev = [
  "ruff>=0.3.5,<0.4.0",
  "black>=24.3.0,<25.0.0",
  "isort>=5.13.2,<5.14.0",
  "pre-commit>=4.0.0,<4.3.0",
  "click>=8.1.0,<8.2.0",
  "h11>=0.14.0,<0.20.0",
  "starlette>=0.46.0,<0.50.0",
  "pydantic>=2.11.0,<2.12.0",
  "anyio>=4.9.0,<5.0.0",
  "cfgv>=3.4.0,<3.5.0",
  "identify>=2.6.9,<2.8.0",
  # "pyyaml>=6.0.0,<7.0.0", # Moved to core dependencies
  # "python-dotenv>=1.0.0,<2.0.0", # Optional: Uncomment to add
  "sqlalchemy[mypy]>=2.0.35,<2.1.0", # For SQLAlchemy 2.0 type hints
  "watchfiles>=0.21.0,<0.22.0",      # Added for worker hot-reloading
  "factory-boy>=3.3.3,<3.4.0",
  "pytest-cov>=6.1.1,<6.2.0",
  "aiosqlite>=0.21.0,<0.22.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 88
target-version = "py311" # Match requires-python

[tool.ruff.lint]
select = ["E", "W", "F", "I", "C", "B"]
ignore = []
fixable = ["ALL"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.black]
line-length = 88
target-version = ['py311'] # Match requires-python

[tool.isort]
profile = "black"
line_length = 88

[tool.setuptools]
packages = ["app"]

[tool.mypy]
python_version = "3.11"                  # Match requires-python
plugins = ["sqlalchemy.ext.mypy.plugin"]
ignore_missing_imports = true            # Optional: Add if needed for other libs
# Add other mypy settings as needed
